+++ index.js
@@ -1,15 +1,21 @@
+const debug = require('debug')('up:index');
+
+const {
+	readFileSync,
+	readdirSync,
+} = require('node:fs');
+const http = require('node:http');
+
 const vgMongo = require('vg-mongo');
+
 const yaml = require('js-yaml');
-const fs = require('fs');
 const lodash = require('lodash');
-const path = require('path');
+const path = require('node:path');
 const bodyParser = require('body-parser');
 const express = require('express');
 const compression = require('compression');
 const cors = require('cors');
-const debug = require('debug')('universal-pattern');
 
-const paginate = require('./libs/paginate');
 const services = require('./services');
 const controllers = require('./controllers');
 const swaggerMetadata = require('./libs/swagger-metadata');
@@ -24,7 +30,7 @@ const getModule = (url) => url.replace('/', '')
 	.shift();
 
 const addHook = (UP) => (endpoint, method, cb) => {
-	debug('adding hook: ', endpoint, method);
+	debug('adding hook: ', endpoint);
 	if (!UP.hooks[endpoint]) UP.hooks[endpoint] = {};
 	UP.hooks[endpoint][method] = cb;
 };
@@ -35,13 +41,16 @@ const registerController = (UP) => (name, controller) => {
 	UP.controllers[name] = controller;
 };
 
-const universalPattern = (app = express(), options = {}) => {
+async function universalPattern(options = {}) {
+	const app = express();
+	const server = http.createServer(app);
+	const folder = path.join(__dirname, './swagger');
 	const defaultOptions = {
 		swagger: {
 			baseDoc: '/core',
 			host: 'localhost',
 			apiDocs: 'api-docs',
-			folder: path.join(__dirname, './swagger'),
+			folder,
 			info: {
 				version: 2.0,
 				title: 'Server API',
@@ -55,30 +64,32 @@ const universalPattern = (app = express(), options = {}) => {
 				},
 			},
 		},
-		compress: false,
-		cors: false,
+		compress: true,
+		cors: true,
 		production: true,
 		routeController: (req, res, next) => { next(); },
+		port: options?.port || 3000,
 	};
 
 	if (options.database) {
-		defaultOptions.database = {
-			uri: 'mongodb://localhost:27017',
-			name: 'upexample',
-		};
+		debug('database setted ');
+		if (!options?.database?.uri || options?.database?.uri === '') throw new Error('Invalid database URI');
+		if (!options?.database?.name || options?.database?.name === '') throw new Error('Invalid database name');
+	}
+
+	if (options.database) {
+		debug('database connected');
+		db = await vgMongo(options.database.uri, options.database.name);
 	}
 
 	const localOptions = lodash.merge(defaultOptions, options);
-
-	if (localOptions.database) {
-		db = vgMongo(localOptions.database.uri, localOptions.database.name);
-	}
 
 	const UP = {
 		localOptions,
 		db,
 		app,
 		getModule,
+		instanceId: new Date().getTime(),
 	};
 
 	app.use(bodyParser.urlencoded({ extended: false }));
@@ -87,38 +98,37 @@ const universalPattern = (app = express(), options = {}) => {
 	if (!hasMWS(app, 'compression') && localOptions.compress) app.use(compression({ level: 9 }));
 	if (!hasMWS(app, 'cors') && localOptions.compress) app.use(cors());
 
-	return new Promise((resolve) => {
-		const yamlContent = [...fs.readdirSync(localOptions.swagger.folder)
-			.map((file) => {
-				if (file.split('.').pop() === 'yaml') {
-					return fs.readFileSync(path.join(localOptions.swagger.folder, file), 'utf8').toString();
-				}
-				return false;
-			}), fs.readFileSync(path.join(__dirname, './swagger', 'index.yaml'), 'utf8').toString()]
-			.map((file) => yaml.load(file))
-			.reduce((acc, current) => lodash.merge(acc, current), {});
+	const yamlContent = [...readdirSync(localOptions.swagger.folder)
+		.map((file) => {
+			if (file.split('.').pop() === 'yaml') {
+				return readFileSync(path.join(localOptions.swagger.folder, file), 'utf8').toString();
+			}
+			return false;
+		}), readFileSync(path.join(__dirname, './swagger', 'index.yaml'), 'utf8').toString()]
+		.map((file) => yaml.load(file))
+		.reduce((acc, current) => lodash.merge(acc, current), {});
 
-		UP.swagger = lodash.merge({ ...yamlContent }, localOptions.swagger, { basePath: localOptions.swagger.baseDoc });
+	UP.swagger = lodash.merge({ ...yamlContent }, localOptions.swagger, { basePath: localOptions.swagger.baseDoc });
 
-		if (!localOptions.production) {
-			app.use(`${localOptions.swagger.baseDoc}/docs/`, express.static(`${__dirname}/libs/swagger-ui`));
-			app.use(`${localOptions.swagger.baseDoc}/docs/`, (req, res) => {
-				const content = fs.readFileSync(`${__dirname}/libs/swagger-ui/index.tmp`).toString();
-				res.end(content.replace('[[URL]]', `${localOptions.swagger.baseDoc}/api-docs`));
-			});
-			app.use(`${localOptions.swagger.baseDoc}/api-docs`, (req, res) => res.json(UP.swagger));
-		}
+	if (!localOptions.production) {
+		app.use(`${localOptions.swagger.baseDoc}/docs/`, express.static(`${__dirname}/libs/swagger-ui`));
+		app.use(`${localOptions.swagger.baseDoc}/docs/`, (req, res) => {
+			const content = readFileSync(`${__dirname}/libs/swagger-ui/index.tmp`).toString();
+			res.end(content.replace('[[URL]]', `${localOptions.swagger.baseDoc}/api-docs`));
+		});
+		app.use(`${localOptions.swagger.baseDoc}/api-docs`, (req, res) => res.json(UP.swagger));
+	}
 
-		app.use(swaggerMetadata(UP));
-		paginate(UP);
-		UP.services = services(UP);
-		UP.controllers = controllers(UP);
-		UP.hooks = {};
-		swaggerRouter(UP);
-		UP.addHook = addHook(UP);
-		UP.registerController = registerController(UP);
-		resolve(UP);
-	});
-};
+	app.use(swaggerMetadata(UP));
+	UP.services = services(UP);
+	UP.controllers = controllers(UP);
+	UP.hooks = {};
+	swaggerRouter(UP);
+	UP.addHook = addHook(UP);
+	UP.registerController = registerController(UP);
+	server.listen(defaultOptions.port);
+	console.info(`ready on *:${defaultOptions.port}`);
+	return UP;
+}
 
 module.exports = universalPattern;
