+++ libs/validators/index.js
@@ -1,0 +1,78 @@
+const debug = require('debug')('up:libs:validators');
+
+const validString = require('./strings');
+const validObject = require('./objects');
+const validNumber = require('./numbers');
+const validBoolean = require('./booleans');
+const validArray = require('./arrays');
+
+const validateParameters = (req, params, level = {}) => {
+	debug('validateParameters called');
+	Object.entries(params)
+		.forEach(([k, v]) => {
+			debug('validating: ', k, v);
+			try {
+				let method = v.in === 'header' ? 'headers' : v.in || 'body';
+				if (v.schema) {
+					method = 'body';
+				}
+				if (v.schema && v.schema.type === 'object') {
+					level[k] = { value: {} };
+					return validateParameters(req, v.schema.properties, level[k].value);
+				}
+
+				if (v.type === 'number' || v.type === 'integer') {
+					const value = validNumber(req, method, k, v);
+					level = { ...level, [k]: value };
+					return level;
+				}
+
+				if (v.type === 'string') {
+					const value = validString(req, method, k, v);
+					level = { ...level, [k]: value };
+					return level;
+				}
+
+				if (v.type === 'boolean') {
+					const value = validBoolean(req, method, k, v);
+					level = { ...level, [k]: value };
+					return level;
+				}
+
+				if (v.type === 'array') {
+					const value = validArray(req, method, k, v);
+					level = { ...level, [k]: value };
+					return level;
+				}
+
+				if (v.type === 'object') {
+					const value = validObject(req, method, k, v);
+					if (value) {
+						level = { ...level, [k]: value };
+						return level;
+					}
+				}
+
+				if (v.type === 'file') {
+					// Object.assign(level, { [k]: { value } });
+					return level;
+				}
+
+				level = { ...level, [k]: req[method][k] };
+				return level;
+			} catch (err) {
+				throw Error(`${err.toString()}`);
+			}
+		});
+
+	return level;
+};
+
+module.exports = {
+	validString,
+	validObject,
+	validNumber,
+	validBoolean,
+	validArray,
+	validateParameters,
+};
