+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "mocha"
+    "test": "node"
   },
   "repository": {
     "type": "git",
@@ -33,20 +33,19 @@
   ],
   "homepage": "https://github.com/lortmorris/universal-pattern#readme",
   "dependencies": {
-    "body-parser": "^1.20.1",
+    "body-parser": "^1.20.2",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "debug": "^4.3.4",
     "express": "^4.18.2",
     "js-yaml": "^4.1.0",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "vg-mongo": "^2.0.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.22.10",
-    "dotenv": "^16.3.1",
-    "eslint": "^8.47.0",
+    "@babel/eslint-parser": "^7.23.3",
+    "eslint": "^8.56.0",
     "eslint-config-airbnb": "^19.0.4",
-    "nodemon": "^2.0.20",
-    "supertest": "^6.3.3"
+    "nodemon": "^3.0.3"
   }
 }
