+++ swagger/index.yaml
@@ -1,6 +1,6 @@
 swagger: "2.0"
 info:
-  version: 0.0.4
+  version: 2.0.0
   title: Universal Patter Server test
   termsOfService: 'http://localhost:8000/terms'
 host: 'localhost'
@@ -76,10 +76,4 @@ parameters:
     in: header
     type: string
     description: bearer token
-    required: true
-  tenant:
-    name: tenant
-    in: header
-    type: string
-    description: Tenant name
     required: true
