--- test/.env
@@ -1,5 +1,0 @@
-HOST=localhost
-PORT=5000
-CONNECTION=mongodb://127.0.0.1/uptesting
-DBNAME=uptesting
-BASEPATH=/services
