+++ test/server.sh
@@ -1,0 +1,6 @@
+export HOST=localhost
+export PORT=5000
+export CONNECTION=mongodb://127.0.0.1:27017
+export DBNAME=uptesting
+export BASEPATH=/services
+nodemon index.js
